refactor: unite struct_tags util with main logic
